refactor(epicenter): move workspace-only modules out of shared/
lifecycle.ts, run-extension-factories.ts, and ydoc-keys.ts were in
shared/ from when both Dynamic and Static APIs consumed them. Now that
Dynamic is gone, they're only used by workspace/ code (plus one import
from extensions/sync). Move them to workspace/ where they belong.

Update all import paths accordingly.
