+++ packages/epicenter/src/workspace/index.ts
@@ -86,15 +86,15 @@ export type {
 	Extension,
 	Lifecycle,
 	MaybePromise,
-} from '../shared/lifecycle.js';
+} from './lifecycle.js';
 
 // ════════════════════════════════════════════════════════════════════════════
 // Y.DOC STORAGE KEYS
 // ════════════════════════════════════════════════════════════════════════════
 
-export type { KvKey, TableKey as TableKeyType } from '../shared/ydoc-keys.js';
+export type { KvKey, TableKey as TableKeyType } from './ydoc-keys.js';
 // Y.Doc array key conventions (for direct Y.Doc access / custom providers)
-export { KV_KEY, TableKey } from '../shared/ydoc-keys.js';
+export { KV_KEY, TableKey } from './ydoc-keys.js';
 
 // ════════════════════════════════════════════════════════════════════════════
 // Schema Definitions (Pure)
